Update to use C++ integer data types defines.

+++ trunk/include/libkea/KEAAttributeTable.h
@@ -55,7 +55,7 @@ namespace libkea{
     {
         size_t fid;
         std::vector<bool> *boolFields;
-        std::vector<long> *intFields;
+        std::vector<int64_t> *intFields;
         std::vector<double> *floatFields;
         std::vector<std::string> *strFields;
         std::vector<size_t> *neighbours;
@@ -105,39 +105,39 @@ namespace libkea{
         virtual KEAATTType getKEAATTType() const;
         
         virtual bool getBoolField(size_t fid, const std::string &name) const throw(KEAATTException)=0;
-        virtual long getIntField(size_t fid, const std::string &name) const throw(KEAATTException)=0;
+        virtual int64_t getIntField(size_t fid, const std::string &name) const throw(KEAATTException)=0;
         virtual double getFloatField(size_t fid, const std::string &name) const throw(KEAATTException)=0;
         virtual std::string getStringField(size_t fid, const std::string &name) const throw(KEAATTException)=0;
         
         virtual void setBoolField(size_t fid, const std::string &name, bool value) throw(KEAATTException)=0;
-        virtual void setIntField(size_t fid, const std::string &name, long value) throw(KEAATTException)=0;
+        virtual void setIntField(size_t fid, const std::string &name, int64_t value) throw(KEAATTException)=0;
         virtual void setFloatField(size_t fid, const std::string &name, double value) throw(KEAATTException)=0;
         virtual void setStringField(size_t fid, const std::string &name, const std::string &value) throw(KEAATTException)=0;
         
         virtual void setBoolValue(const std::string &name, bool value) throw(KEAATTException);
-        virtual void setIntValue(const std::string &name, long value) throw(KEAATTException);
+        virtual void setIntValue(const std::string &name, int64_t value) throw(KEAATTException);
         virtual void setFloatValue(const std::string &name, double value) throw(KEAATTException);
         virtual void setStringValue(const std::string &name, const std::string &value) throw(KEAATTException);
         
         virtual bool getBoolField(size_t fid, size_t colIdx) const throw(KEAATTException)=0;
-        virtual long getIntField(size_t fid, size_t colIdx) const throw(KEAATTException)=0;
+        virtual int64_t getIntField(size_t fid, size_t colIdx) const throw(KEAATTException)=0;
         virtual double getFloatField(size_t fid, size_t colIdx) const throw(KEAATTException)=0;
         virtual std::string getStringField(size_t fid, size_t colIdx) const throw(KEAATTException)=0;
         
         virtual void setBoolField(size_t fid, size_t colIdx, bool value) throw(KEAATTException)=0;
-        virtual void setIntField(size_t fid, size_t colIdx, long value) throw(KEAATTException)=0;
+        virtual void setIntField(size_t fid, size_t colIdx, int64_t value) throw(KEAATTException)=0;
         virtual void setFloatField(size_t fid, size_t colIdx, double value) throw(KEAATTException)=0;
         virtual void setStringField(size_t fid, size_t colIdx, const std::string &value) throw(KEAATTException)=0;
         
         virtual void setBoolValue(size_t colIdx, bool value) throw(KEAATTException);
-        virtual void setIntValue(size_t colIdx, long value) throw(KEAATTException);
+        virtual void setIntValue(size_t colIdx, int64_t value) throw(KEAATTException);
         virtual void setFloatValue(size_t colIdx, double value) throw(KEAATTException);
         virtual void setStringValue(size_t colIdx, const std::string &value) throw(KEAATTException);
         
         virtual KEAATTFeature* getFeature(size_t fid) const throw(KEAATTException)=0;
         
         virtual void addAttBoolField(const std::string &name, bool val, std::string usage="") throw(KEAATTException);
-        virtual void addAttIntField(const std::string &name, long val, std::string usage="") throw(KEAATTException);
+        virtual void addAttIntField(const std::string &name, int64_t val, std::string usage="") throw(KEAATTException);
         virtual void addAttFloatField(const std::string &name, double val, std::string usage="") throw(KEAATTException);
         virtual void addAttStringField(const std::string &name, const std::string &val, std::string usage="") throw(KEAATTException);
         virtual void addFields(std::vector<KEAATTField*> *inFields) throw(KEAATTException);
@@ -171,7 +171,7 @@ namespace libkea{
         static H5::CompType* createKeaStringCompTypeDisk() throw(KEAATTException);
         static H5::CompType* createKeaStringCompTypeMem() throw(KEAATTException);
         virtual void addAttBoolField(KEAATTField field, bool val) throw(KEAATTException)=0;
-        virtual void addAttIntField(KEAATTField field, int val) throw(KEAATTException)=0;
+        virtual void addAttIntField(KEAATTField field, int64_t val) throw(KEAATTException)=0;
         virtual void addAttFloatField(KEAATTField field, float val) throw(KEAATTException)=0;
         virtual void addAttStringField(KEAATTField field, const std::string &val) throw(KEAATTException)=0;
         virtual KEAATTFeature* createKeaFeature() throw(KEAATTException);
